I have finished with download and delete methods in repository. I added default image to personal card

+++ src/main/java/ua/in/gnatyuk/repository/GoogleDrivePhotoRepositoryImpl.java
@@ -12,6 +12,7 @@ import ua.in.gnatyuk.utils.MyGoogleDrive;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.Optional;
 
 @Component
 public class GoogleDrivePhotoRepositoryImpl implements PhotoRepository {
@@ -43,23 +44,40 @@ public class GoogleDrivePhotoRepositoryImpl implements PhotoRepository {
     public boolean delete(String photosId) {
         try {
             Drive drive = googleDrive.getDriveService();
-            drive.files().delete(photosId).execute();
-            return true;
+            File file = drive.files().get(photosId).execute();
+            if (!file.isEmpty()) {
+                drive.files().delete(photosId).execute();
+                return true;
+            }
         } catch (IOException e) {
             e.printStackTrace();
-            return false;
         }
+        return false;
     }
 
     @Override
     public java.io.File download(String photosId) {
         try {
             Drive drive = googleDrive.getDriveService();
-            java.io.File tempFileForPhoto = new java.io.File("/temp/");
+            java.io.File tempFileForPhoto = java.io.File.createTempFile("photo","tmp");
             FileOutputStream fileOutputStream = new FileOutputStream(tempFileForPhoto);
             drive.files().get(photosId).executeMediaAndDownloadTo(fileOutputStream);
 
             return tempFileForPhoto;
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    public String findIDByFileName(String fileName) {
+        try {
+            Drive drive = googleDrive.getDriveService();
+            Optional<File> resultOfSearch = drive.files().list().execute().getItems().stream().filter(e->e.getTitle().equals(fileName)).findFirst();
+            if(resultOfSearch.isPresent()){
+                return resultOfSearch.get().getId();
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
